Revert "feat: :sparkles: Handle konnectors with AppCollection"
This reverts commit 6c35e009fd1c25975cfe58744a93fe292eccd229.
This commit is problematic because it introduces a breaking change.
Before we use `io.cozy.konnector` from `/data/io.cozy.konnector` and
new we used `/konnector`.
On cozy-banks we use `io.cozy.konnector` with /data stack route
to use [mongo query][query].
After discussion with @CPatchane, we decide to revert this commit
because it's not necessary for now. But it's not a good practice to go
through this path so we will have to redo this PR later.

[query]: https://git.io/fjeVg