feat(cozy-client): Expose main and browsers entry points
The main entry point is exposing the API without react bindings. This
allows a node application to use cozy-client without having to import
modules from the `dist` directory and without having to install react as
a dependency.

The browsers entry point is exposing the entire API (with react
bindings). This changes nothing for web apps.

There is also the possibility to import react specific stuff from
`cozy-client/dist/react` in rare cases in which this could be useful.
This is just an escape hatch.