fix: Require regenerator-runtime from @babel/runtime/regenerator
At some point cozy-client was shipped on npm with regeneratorRuntime
not being correctly imported. It must be because in 1e53a167fa43117d1027423b814f9000f427da87,
the plugin transform-runtime was no longer imported, and did not override
the transformRegenerator: false specified in the line just above. As a
lib, cozy-client should not make any assumption on its environment, ie
it should not take regeneratorRuntime from global/root