feat(cozy-client): Add getStoreURL method to applications model
With getStoreInstallationURL, the user was redirected to the store page
with the permissions modal. We want to be able to redirect to the app's
store page and make the user click on the "install" button, otherwise he
may not know what he is going to install. The getStoreURL method returns
the app's root page. And getStoreInstallationURL now uses it to return
the installation URL.