test: Use mount instead of shallow
shallow()/dive() makes an assumption on the structure how the tested
component. It breaks easily, for example when Provider changes its
structure. Moreover, bugs in enzyme force us to provide the context,
which meant the test was not really testing the providing of the context.
With mount, we do not have any of those problems