fix: Remove deleted document from store
Previously, deleted documents were removed from queries, but the doc
was staying in the documents slice. This is problematic in the case
of relationships since the "parent" document linked to a document
that has been deleted, will still see the deleted document.

This is also problematic if we simply rely on the documents slice to
know if a document is deleted or not.

This commit fixes both situations.