fix: Do not set loading on queryState for noninitial initQuery
Previously, after an initial load of data through query(), subsequent
calls to query() would set the fetchStatus to loading. It causes a
problem in case of fetchPolicy : if the fetchPolicy returns false, the receiveQueryResult will never been dispatched and queryState.fetchStatus
will stay to false.

To prevent the problem, an additional loadQuery call is introduced which
sets the fetchStatus to loading. initQuery now does not set the
fetchStatus to loading if the query already has been fetched.