feat: When an app/konnector is fetched by id, doctype should be present
This is not breaking but this is deprecating the get with only the slug.

Q('io.cozy.apps').getById('banks') // Deprecated
Q('io.cozy.apps').getById('io.cozy.apps/banks') // OK

This is because otherwise we have an erroneous behavior when loading
all konnectors then only one.

client.query(Q('io.cozy.konnectors').getById('konn'), { as: 'konn' })
client.query(Q('io.cozy.konnectors'), { as: 'konnectors' })

At this point "konn" query will have nothing in its data since it was
cleared out by the filtering that cozy-client does when receiving new
documents, since the id "io.cozy.konnectors/konn" does not correspond
to the id in the queryDefinition "konn".

To prevent this behavior, we have to include the full id when doing the
query.
To prevent a breaking change, the old behavior is called out as
deprecated via a console.warn.