feat: Optimistic index creation
We used to have a pessimistic index creation strategy: for each find
query, a index creation query was sent to the database, and ignored if
the index was already created.
Most of the time, these index creation queries are useless, as the index
is created when the app is first used.
Therefore, we now adopt an optimistic strategy, that is, running the
find query, and create the index afterwards only if a missing index is
returned.
It also now creates a named index, i.e. `_design/by_selector_field` and
catch index creation conflicts.