feat: Copy design doc when equivalent index is found
With the optimistic index creation approach, we are going to create new
design documents, as the old ones used a CouchDB-generated ID, while
the new ones will have a deterministic name, e.g. `by-date`.
It would be a pity to re-compute all existing
indexes, so, to avoid this, we check if there is an equivalent index
definition in the design docs, and copy it, if it is the case.
This copy will use the same index and won't need to recompute
everything.