feat: Limit to 1 the queries after index creation
We force a query on newly created index in order to ensure they are
actually created, as we experienced situations in production where a
created index would actually be shortly unavailable after creation.
We pass a limit to 1, in order to reduce the query cost, as it is just
used to test the index availability.