feat: Can use app_slug when creating a sharing
By default, Cozy-Stack gets the app_slug from the token.
But we can define the app_slug when creating the sharing. In that case
the cozy-stack will use this slug in the mail sent for a cozy-to-cozy
sharing.
It enables the fact, that the app1, can generate a link to an app2.