fix: Detect and destroy inconsistent indexes
When trying to find a matching index, we used to sort the fields to
perform the comparison.

Therefore, this situation could arise:

1. Create an old index (e.g. _design/123) on fields [foo, bar]
2. Create an old index (e.g. _design/456) on fields [bar, foo]
3. With the new optimist index strategy, if we try to query on [bar,
foo] and the index does not exist, we will try to find an existing
matching index to migrate it as a new index. Here,  _design/123
will match because its sorted fields results in [bar, foo].
4. We migrate the index _design/123 into _design/by_bar_and_foo,
which actually index [foo, bar]. The _design/123 index is deleted.
5. When trying to query on [bar, foo], the _design/by_bar_and_foo index
is used, but couchdb returns an error, as the fields order is incorrect
(it actually indexes [foo, bar])
6. Because of this error, we try again to find an existing index:
_design/456 matches. We try to migrate it but it results into a
conflict because _design/by_bar_and_foo already exists.

So, to solve this, we remove the sort on the fields,
and check if an index is in an inconsistent state,
i.e. if its name does not correspond to the indexed fields order,
typically _design/by_bar_and_foo that indexes [foo, bar]. If
this is the case, we delete it.