feat: Add devtools
- Add a devtool panel that can be opened by clicking on a Fab button
that is only displayed when the debug flag is set to true.
- The devtools is made of several "panels". There are deffault panels
and the app can also inject its own adhoc panels.
- The "queries" panel shows the current queries inside ccozy-client
cache.
It is inspired by react-query devtools and should in the
future be improved to allow for action on queries: reset, set to empty,
delete etc... It could also show performance tips for each query (based
on the definition for example).
- The "flags" panel shows all the current flags and allow to modify them.
- The "libraries" panel show library versions based on the
global __VERSIONS__ variable that should be injected by the app. If it
is defined, the panel will be blank.