feat: Do not fetch the same query is she is already loading
In some case we can have the same query running almost in parallel.
In that case, we juste want only one to be executed. Others will
receive automatically the results.

To do that, I remove a previous optimisation saying that when we
INIT a query we force the fetchStatus to loading. It's true in most
case but not in the case I explained before.

So now, init only init, load only load and we can check that when we
try to load if something is already loading.