feat: Initiate replication with _all_docs + bulk insert
At initialization time, we used to replicate docs from CouchDB to
PouchDB by using the replication protocol. This is problematic
because it replicates the revision history, making the indexing
potentially slower and increasing the database size.
The replication itself is also slower than just queyring the _all_docs
and inserting the documents in bulk.
Here are some numbers we measured on a database with 10K docs and
about 100K updates:

* Replication with CouchDB replication protocol: 34s
  * Includes revision history
* Replication with _all_docs + bulk insert: 13s
  * Does not include revision history
* Indexing with revisions: 38s
* Indexing without revisions: 26s

Note this only concern the initial replication, we continue to use the
CouchDB replication protocol for the next runs.

test: Add tests for replication

fix: Wrap promise