feat: Save last document _id when replicating with all_docs
When doing the first replication, through _all_docs, we insert in bulk
the documents. If the replication cannot perform to the end, we will
have to start again. To avoid this, we save in localStorage the last
saved document _id, so we can start from here next time.
Note the bulk insertion of already inserted docs is not a big deal, as
this will just re-insert them, with no new revision.

chore: Extract bulk insert in helpers

refactor: Remove logs

style: Lint