feat: Queries can choose not to be auto updated
Options passed when a query is executed are stored in the store. This
gives us the ability to check the autoUpdate option and if set to false
to forego the auto updating mechanism.

This is useful for sorted queries on a date interval. If I query
bank transactions for march and then query those for january, I do not
want the transactions from january to go into the march query, otherwise
we can have a bad state since the february transactions are not there.