feat: Include_docs is done by default by fetchChanges
Fetch changes needs to have the docs included in the changes response
since it filters null documents, deleted documents, and design docs.
With include_docs: false, fetchChanges was not functional, thus it makes
no sense to support it. Now, the caller of fetchChanges doesn't need to
pass include_docs: true, it is set automatically by
DocumentCollection::fetchChanges.

Since fetchChanges gets higher level, we added a lower level method: fetchChangesRaw.

It is closer to the _changes API and does not deal with filtering of
design docs and deleted documents.