docs: Generate docs for cozy-client via typedoc
jsdoc does not correctly parse some typescript supported syntaxes
(imports for example, or type property extraction). By using typedoc,
we stay in the typescript ecosystem and use the same parser.

Here, jsdoc is replaced by typedoc to generate the API documentation
of cozy-client in markdown.

remark is used to postprocess the generated markdown
documentation to strip deep headings and replace them with emphasis
nodes so that they do not appear on the right table of contents on
https://docs.cozy.io.

An interesting side effect is that each class / or namespace is on
one page, it makes it easier to navigate the doc and to see the API.