fix: Cached queries optimization
In https://github.com/cozy/cozy-client/pull/925 we added the check on
the fetch status to return nothing when a query was already loading.

In https://github.com/cozy/cozy-client/pull/935 we decided to go deeper
and return the same promise when the query was already loading but
without any early return. The result was a perf degredation in our apps
since CozyClient was dispatching loadQuery / ResultQuery each time. So
if the request was done 10 times, the request was only executed once,
but the dispatch was done 10 times. Resulting in a lot of renders within
the app.

Here, we combine the best of the two PRs. We early return to bypass the
dispatch events, but we also return the stored promise to be able to get
the result even in a world without React.