feat: Adapter migration
PouchDB can use different kind of adapter, e.g. idb, sqlite, etc.
We want to be able to migrate the adapter, typically to use the new and
more performant IndexedDB adapter.

To perform such migration, we run a local replication between the old
and new database instance. By doing so, the user can migrate its data
without requiring an internet connection.

Note the migration can take some time, i.e. several minutes depending on
the device, the database size, etc.
It is the responsability of the app to to call the `migrateAdapter` API
and handle this migration on the UI level.

Once the adapter is migrated, the old database is removed, and the page
reloaded so the client will use the new Pouch. The adapter name is saved
in the localstorage to inform the app which adapter should be used.

fix: Empty adapter