feat: Allow to query a HasManyFiles relation in both ways
When using a HasManyFiles relationship, we can manipulate documents from
the referenced doc, or from the referencing doc.

Let's assume the following photo albums:

```
{
  _id: '1234'
  _type: 'io.cozy.photos.albums'
}
```

And the following file, with a reference to the album:

```
{
  _id: '5678'
  _type: 'io.cozy.files',
  referenced_by: [
    {
      id: '1234'
      type: 'io.cozy.photos.albums'
    }
  ]
}
```

Previously, it was only possible to query the files from the album.
But, we might want to query the album directly from the file. This is
what we add here, by querying documents specified in the `referenced_by`
relationship specified in the file, through their id.

test: Add test for HasManyFiles query