feat: Infer content-type from file extension
We used to infer the file content-type only for File objects. However,
it can be useful to extend this for binary structs such as ArrayBuffer,
typically when dealing with encryption. Say we have decrypted a file,
represented as an ArrayBuffer: we have no way to know its content-type
without inferring it from the name, as it was previously encrypted with
a specific content-type.