feat: Makes queryAll faster with many requests
When there many documents in the same doctype, and queryAll is used to
fetch them all, it can be very slow. It looks like Cozy-Client is taking
more and more time between requests to update its state.

The extractAndMergeDocument function is used to merge the documents in
state and the documents in the response. It was iterating on both sets
of documents, which seems useless. By iterating on just the new
documents, it can be faster and avoid the growing time between requests.