refactor: Use forceFileDownload in file collection
Moving this method in the file collection allows to access it from an
app. This is typically useful for an encrypted file with Drive, because
we do not want the download link to be handled by the stack, as we have
to decrypt the binary first and then generate an URL from it.