fix: Add id in CozyClientDocument for RealTime
Although we tend to rely on the `_id` for all documents, we sometimes
also rely on `id`, which is a legacy that we encounter in some apps such
as Drive. Therefore, e23239fc2a17f8362a74dabdda0e3f816dfc6eb2 introduced
a regression in Drive as the `id` was missing.