fix(documents): Reducer should assign new object for concerned doctype
In Banks, we noticed the reference of the state.cozy.documents[doctype]
does not change when we refetch more transaction.

This was dued to extractAndMergeDocument that created a new
mergedData object but providing again the same properties.

Our property that got updated, from 1000 documents to 1500, was still
inside the same reference, so reselect did not recompute after store
got updated, and users faced this bug while moving from home page with
reimbursements, users see old reimbursements from home/transactions
instead of seeing the full reimbursements fetched on Reimbursements page.