feat: Handle user cancelation from ReactNative's InAppBrowser
When displaying the `/authorize` route using ReactNative's InAppBrowser
the user has two ways to cancel the authorization

First they can touch the InAppBrowser's close button. In this case we
now intercept the event and reject the promise with a USER_CANCELED
message

They can also touch the authorization dialog's close button from the
displayed HTML page. In this case the cozy-stack will do a redirection
with `error=access_denied` in URL query parameters. We intercept this
URL and reject the promise with the same USER_CANCELED message

Related PR: cozy/cozy-stack#3338