feat: Support client.save for files
When updating a file we might either want to update its CouchDB
attributes, e.g. the name, or the binary itself.
To do so, the app used to go through the collection, and call `.update`
for attributes or `.updateFile` to update the binary.
It is now possible to directly use `client.save` for both cases. When a
`data` attribute is passed to the document, is is considered as a binary
update.