feat: Throw an error when fields do not include all searched attributes
BREAKING CHANGE: when a `.select` is used in a query definition, it
MUST includes all the fields expressed in the selector and partial
index.
This is because we evaluate queries on the store thanks to sift,
and it requires to have all the necessary fields to evaluate the
query. If one is missing, the query will return nothing.
See this example:

```
Q('io.cozy.todos).where({date: $gt: { null}}).select(['name'])
```

When this query will be run, only the todos name will be in the store.
Then, when later evaluated in the store, sift won't find any todos with
a date ; thus it will return nothing and consider all the existing todos
should be removed from the store.