fix: Make a correct deep merge on objects.
When we merge object with the spread operator, it
doesn't make a deep merge, it replace the second
level.

Exemple:
```js
d1 = {a: 1, b: {b1: 1}}
d2 =  {c: 3, b: {b2: 2}}
merge = {...d1, ...d2}
// merge = { a: 1, b: { b2: 2 }, c: 3 }
```

This is not what we want in this case. We want a
fully merged object, not a replace. We expect the
result to be:
```
// mege = { a : 1, b { b1: 1, b2: 2}, c: 3}
```

To achieve that, we use lodash merge!

Fix #1175