feat: Handle CouchDB fallback for partial index
When a partial filter is defined and the index is not created yet,
CouchDB might fallback on another existing index and return a warning
saying the specified index was not used.
This is not what we want, because the existing index might not actually
use the partial filter and make a full-scan on the index to evaluate
it.