fix: Mark the query as loaded when the document does not exist
When we query a document by passing by the Collection.get()
we do not throw if we got a not_found. We simply return null
https://github.com/cozy/cozy-client/blob/8ec78eb5a028f8df4908e283d0d48572d475ab3f/packages/cozy-stack-client/src/Collection.js#L103

By returning null is that case, means that
we return the state as is https://github.com/cozy/cozy-client/blob/8ec78eb5a028f8df4908e283d0d48572d475ab3f/packages/cozy-client/src/store/queries.js#L167-L169
because data is null.

Since it returns the state as is, if this is
the first time we run the query, the state is
blocked to fetchStatus === loading.

Since the fetchStatus is still in loading state
then if we re-run the request, then CC will ignore
it by its deduplication mecanisme https://github.com/cozy/cozy-client/blob/master/packages/cozy-client/src/CozyClient.js#L922-L925

So we never get the right result.

The fix is to change the state when there is
no result. We change the fetchStatus but also
the lastFetched and updated attribute.