feat: Fetch capabilities on OAuth login
Current cozy-client implementation calls `loadInstanceOptionsFromDOM`
when initialized

This method only works for web app clients, but would fail for OAuth
clients as those cannot rely on any DOM

To fix this we should retrieve client's options by calling the
cozy-stack

For now we only need to retrieve capabilities. But other props should
be implemented later

As this method is async, we don't want to call it from the constructor.
Instead we chose to call it at the end of `login()`. Which would also
allow to access token data when we will implement this part