fix: Limit DocsIds according to limit instead of queryState.count
Bugs happen when creating a shortcut, inside a new folder in Drive
queryState returns a data as an empty array, so count was 0
And sliceCount removes the new shortcut ids create, because it is 0
The fix using a limit - to 100 in our case, handles that case.
And should not break other scenarii.