fix: Remove isReferencedByAlbum
BREAKING CHANGE: Since the introduction of the photos' clustering, each photo is at least referenced by an automatic album. Therefore, this method which was used to know if a file was referenced by an album, will always return true, does not make sense anymore because it is not what we want to test.

Instead, we now want to see if a file is not referenced by an album that has not been created automatically. This implementation has been done in Drive directly in eb76c831b3deabd0857b9472451830831df5fb3a