fix: Do not call the queryDefinition if enabled is false
It's a legit use case to have something like this:
```
const id = useParam('id');
const file = useQuery(Q('io.cozy.files').getById(id),
{
enabled: !!id
}
if(!id) return null
if(file.fetchStatus === 'loaded'){
   return <File file={file.data} />
}
...
```

This case was failing beforce because getById() checks if
the `id` is there or not. And since even if `enabled` was
set to false we called the getById(), it throws an error.

This fix a part of https://github.com/cozy/cozy-client/issues/961

I don't know yet what to think about the `fetchStatus` equal to
`pending` or `loaded` so I just added a test to be sure to create
a BC if we change the behavior. But to me, the called should not
rely on the result of the useQuery hook when enabled is set to
false.

This fix also enable the fact to be able to call useQuery without
a queryDefinition if `enabled` is set to false. I don't like it
but it's a collateral damage.
The difference between this PR and this one
 https://github.com/cozy/cozy-client/pull/970

is that this is not the responsability to cozy-client to handle
the result of the enabled props. Since `enabled` is passed by
the caller, the caller has the responsability to not use the
data coming from the useQuery hook. Like I did in the begining
of this commit.