feat: Add expiration helpers
This extracts and rewrites expiration helpers from `mespapiers` that will be used directly in `cozy-ui`, `cozy-mespapiers-lib` and also `mespapiers`.

This avoids using `paperDefinitions` as a config.
It may not be generic enough in the future and
we don't have any need for that right now.

This also removes the need to pass a `dateLabel` argument to these helpers.

This also adds new helpers, one giving the notice link if available, and two other for comparing dates.

This also removes the default notice period, meaning that papers won't be considered as expiring if the user has not set any notice period and that the alert won't appear and no notification will be sent.