fix: Do not modify selector with lodash merge
Lodash merge modify the source object to apply the merge
on it.
That was causing errors on queries without `.indexfields` but with
selector and a partial filter.
In such situation, the final selector is a merge between the actual
selector and the partial index.
After the first query, the selector became the merge one.
Then, the query is replayed when the index is missing, this time with
a merged selector.
As the indexed fields were not specified, and we are using the
selector fields to guess them, the resulting index was wrong, as
partially based on the partial filter.