fix: Normalize doc for fetchChanges
Recently, we changed in Banks from
BankTransaction.updateAll() to client.saveAll().

But we had errors in our log saying that we were
calling collection without a doctype
(CozyStackClient.collection() called without a doctype)

In fact, this is because saveAll need to read the
_type for the documents to update them cf:
https://github.com/cozy/cozy-client/blob/master/packages/cozy-client/src/StackLink.js#L122

Sometime it worked for Banks sometime not. It was due
to the fact that we have a condition here:
https://github.com/cozy/cozy-banks/blob/master/src/targets/services/helpers/helpers.js#L19-L45

The fix, is to normalize the document returned by fetchChanges()
like that we add _type & id automatically.

We have a fetchChangesRaw() to get the real changes without any
edition.