fix: Use id instead of path in SettingsCollection.get()
Previously, get was receiveing a path like "instance". This path
was used to call a correct route like "/settings/instance" but was
also used internally as an id for store stuff. But the id
of the document behind "/settings/instance" is
"io.cozy.settings.instance" and not "instance" leading to store issue.