feat: Expose setAppMetadata
In the context of the flagship app, we create a new CozyClient
for the clisk. We set the appMetadata we have when we do the
instanciation aka slug & version. But in some scenario, we
don't have yet the accountIdentifier. So we can't pass it at
this moment.

In order to still write this metadata, we now expose a
setAppMetadata() to be able to use it later after the creation
of the client.