feat: Do not update the document's slide if the doc is up to date
The idea is to avoid unecessary update of the document's slice
when the doc is alreay up to date. In that case, if we already
have the same id in the store, with the same revision, then we
skip the update in order to keep the same object reference to
avoid uncesseray re-render on the app side.

This scenario can be tested by this one:
- App call saveAll(newDocs)
- Cozy-Client call bulk_docs
- Cozy-Client update its document slice with the newDocs and
add the new _rev to it
- App received real time event for created / updated document
- App call new mutation result with the same _rev as before
and with the same content as newDocs[0]

Now, Cozy-Client will not update its store and avoid re-render.