fix: Do not call refreshToken several times in parallel
We can have several fetchJSON() in parallel and if the token
is old, then we can call several times refreshToken.

This refresh token mechanism update the io.cozy.oauth-client
document and can lead to couchdb conflict.

In order to avoid this issue, we store the first promise and
use this one for other call. Like that, the refreshToken
method is only called once.

I copy / paster the promise-cache file, because I don't want to
include a dep from cozy-client inside stack-client. There is
already issue talking about that.