fix: Stop destroying inconsistent index
We have two issues with indexes in Cozy-Client:

The first one is related to the usage of logical mango operators ($or,
$and, $in, etc.) at the top level.Cozy-Client doesn't handle them
properly and creates indexes named $or/$and/$etc
(see https://github.com/cozy/cozy-client/issues/1270). This can lead
to conflicts with the indexes.

Even worse, when we create a partialFilter without declaring a
top-level $and, like [{foo: bar, truc:muche}], the index name on the
Cozy-Client side is correct (`foo_truc`). However, when we inspect
the design doc, CouchDB automatically adds the $and, and it becomes
`$and: [{foo: bar}, {truc: muche}]`

The second issue is the automatic deletion of what Cozy-Client
considers "inconsistent" indexes. When Cozy-Client receives an event
of type `index_not_found` or `not_usable_index`, it performs some
leanup by querying all the indexes on the relevant doctype and checks
if the index name matches the indexed fields in the design doc. Due
to CouchDB's automatic addition of a top-level logical operator with
an implicit $and, Cozy-Client doesn't handle this well and considers
the index as "inconsistent," leading to its deletion.

The real solution is to make Cozy-Client capable of handling logical
operators at the top level of indexes / partialFilters. To achieve
this, we need to rename all the indexes. However, doing so would
result in the deletion of indexes judged inconsistent by the current
Cozy-Client version. Newer versions of Cozy-Client will create new,
correct indexes, but older apps will delete them as they won't
correspond to the previous logic. This process of create / delete /
create / delete will become a nightmare.

Therefore, before changing the way we name indexes, we will remove
the deletion of indexes judged "inconsistent." This simplifies the
app update process. Once the apps are updated with this modification,
we can proceed with renaming the indexes.

We will no longer perform cleanup via Cozy-Client in the future and
will use ad-hoc scripts for that purpose.

This commit represents the first step: we stop deleting "inconsistent"
indexes.