fix: Magic Folders localization does not work with Metro
Metro does not support dynamic requires and the error was swallowed by
an empty catch.

Replaced the dynamic requires with explicit imports.

I missed this problem before because the warning was not obvious in
flagship app logs and since I did not remove magic folders for my tests
but just the destination folder, magic folders references where still
present and did their job.