feat: Benefit from the cache for the apps in maintenance
Since cozy-client 41.9.0 release, we can benefit from the cache for the request to get the apps in maintenance using Q('io.cozy.apps_registry').getById('maintenance'). This commit updates the useAppsInMaintenance hook to use this request with useQuery.

BREAKING CHANGE: useAppsInMaintenance no longer requires a cozy-client instance but to be placed within a CozyProvider