fix: Handle maintenance app without id
There are two ways of maintaining an app using registry and cozy-stack.  Apps from the registry have an id that corresponds to the slug. Conversely, apps from the stack do not have an id, which causes cozy-client to crash. I've decided to create an id from the slug for those that don't have one to fix the problem temporarily.