feat: Apply data enhancement to realtime results for specific doctypes
This enrichment closes the gap between the results of queries from the
cozy-stack and realtime. In fact, realtime returns the document
directly from the database, unlike the stack, which can add computed
fields. Without them, it can lead to instable behaviour in
applications. Some applications fix some issue by reimplementing the
RealTimeQueries component. This commit aims to provide unifed api to
avoid duplicated code. It can also be used as a reference when we want
to close the gap.