fix: Get shortcut img for links from applications in the same cozy
Cozy applications only make their assets available if you are authenticated. As a result, the stack cannot retrieve the favicon. When a link comes from an application in the same Cozy, we can find out its slug and deduce its icon. For the moment, we can't do this for third-party Cozy applications because the URL formats are too diverse.